requirements.txt
flask
flask_cors
pyyaml
gensim
OpenNMT-tf
Pillow
scipy
numpy